% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predators.R
\name{aggregate_predators}
\alias{aggregate_predators}
\title{Aggregate predator sightings}
\usage{
aggregate_predators(predators)
}
\arguments{
\item{predators}{Predator data (from raw CSV file)}
}
\value{
Predators aggregated in intervals
}
\description{
Aggregate predator sightings
}
