% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{ant_proj}
\alias{ant_proj}
\title{Use WGS 84 / Antarctic Polar Stereographic for mapping, recentered on ant_lims()}
\usage{
ant_proj()
}
\value{
CRS object
}
\description{
Use WGS 84 / Antarctic Polar Stereographic for mapping, recentered on ant_lims()
}
