% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predators.R
\name{assign_sightings}
\alias{assign_sightings}
\title{Assign sightings to nearest station}
\usage{
assign_sightings(sightings, stations, max_dist_km, max_days)
}
\arguments{
\item{sightings}{predator sighting data}

\item{stations}{station data}

\item{max_dist_km}{maximum spatial tolerance (km)}

\item{max_days}{maximum temporal tolerance (days)}
}
\value{
predator data associated with nearest station
}
\description{
Assign sightings to nearest station
}
