% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice.R
\name{clean_ice}
\alias{clean_ice}
\title{Clean ice categories}
\usage{
clean_ice(ice_types)
}
\arguments{
\item{ice_types}{Character vector of ice types}
}
\value{
Cleaned up ice types
}
\description{
Converts e.g., FY/MY to FY and collapses SLUSH and BRASH into THIN
}
