% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predators.R
\name{code_to_common}
\alias{code_to_common}
\title{Convert species code to common name}
\usage{
code_to_common(species_code)
}
\arguments{
\item{species_code}{4-letter species codes (character vector)}
}
\value{
Species' common names
}
\description{
Convert species code to common name
}
