% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predators.R
\name{code_to_scientific}
\alias{code_to_scientific}
\title{Convert species code to scientific name}
\usage{
code_to_scientific(species_code)
}
\arguments{
\item{species_code}{4-letter species codes (character vector)}
}
\value{
Species' scientific names
}
\description{
Convert species code to scientific name
}
