% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice.R
\name{collect_seaice}
\alias{collect_seaice}
\title{Collect sea ice raster data}
\usage{
collect_seaice(seaice_dir)
}
\arguments{
\item{seaice_dir}{directory with August sea ice concentration}
}
\value{
terra::SpatRaster with sea ice concentration data
}
\description{
Collect sea ice raster data
}
