% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{create_ant_sf}
\alias{create_ant_sf}
\title{Create Antarctic Peninsula sf object using Natural Earth data}
\usage{
create_ant_sf(ne_dir, limits = ant_lims())
}
\arguments{
\item{ne_dir}{directory to store Natural Earth data}

\item{limits}{spatial limits}
}
\value{
Antarctic Peninsula sf object
}
\description{
Create Antarctic Peninsula sf object using Natural Earth data
}
