% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmds.R
\name{env_data}
\alias{env_data}
\title{Combine ice and station environmental data}
\usage{
env_data(stations, ice, predators)
}
\arguments{
\item{stations}{Zooplankton station data}

\item{ice}{Ice data}

\item{predators}{Predator data}
}
\value{
Environmental data for NMDS loadings
}
\description{
Filters to stations used in this analysis
}
