% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{latlon_to_sf}
\alias{latlon_to_sf}
\title{Convert latitude/longitude data to sf object}
\usage{
latlon_to_sf(df, coords = c("x", "y"))
}
\arguments{
\item{df}{Data frame with spatial data}

\item{coords}{Names of columns containing longitude and latitude}
}
\value{
df converted to sf object
}
\description{
Convert latitude/longitude data to sf object
}
