% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_clustenvtbl}
\alias{make_clustenvtbl}
\title{Create cluster x environment table}
\usage{
make_clustenvtbl(stations_clust)
}
\arguments{
\item{stations_clust}{Station-level cluster data}
}
\value{
knitr::kable table
}
\description{
Create cluster x environment table
}
