% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_commfreqtbl}
\alias{make_commfreqtbl}
\title{Create predator community frequency table}
\usage{
make_commfreqtbl(stations_clust)
}
\arguments{
\item{stations_clust}{Station-level cluster data}
}
\value{
knitr::kable table
}
\description{
Note: originally I implemented this using flextable output, but there's a bug
involving flextable, docx output, and table captions. See
https://github.com/quarto-dev/quarto-cli/issues/9922.
}
