% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{make_fig_nmds}
\alias{make_fig_nmds}
\title{Create a figure with NMDS loadings
years}
\usage{
make_fig_nmds(nmds_envfit, nmds_df, nmds_sightings, nmds_env)
}
\value{
ggplot object
}
\description{
Create a figure with NMDS loadings
years
}
