% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{make_fig_predclustannual}
\alias{make_fig_predclustannual}
\title{Create a figure with geographic distribution of predator clusters by year}
\usage{
make_fig_predclustannual(stations_clust)
}
\value{
ggplot object
}
\description{
Create a figure with geographic distribution of predator clusters by year
}
