% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{make_fig_predclustkde}
\alias{make_fig_predclustkde}
\title{Create a figure with geographic distribution of predator clusters across
years}
\usage{
make_fig_predclustkde(stations_clust)
}
\value{
ggplot object
}
\description{
Create a figure with geographic distribution of predator clusters across
years
}
