% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{make_fig_seaiceclust}
\alias{make_fig_seaiceclust}
\title{Create a figure with distribution of sea ice by predator cluster}
\usage{
make_fig_seaiceclust(stations_clust)
}
\value{
ggplot object
}
\description{
Create a figure with distribution of sea ice by predator cluster
}
