% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{make_fig_top6map}
\alias{make_fig_top6map}
\title{Create a figure with climatological distribution of six most abundant
predators}
\usage{
make_fig_top6map(topsix_clim_class)
}
\value{
ggplot object
}
\description{
Create a figure with climatological distribution of six most abundant
predators
}
