% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_indicatortbl}
\alias{make_indicatortbl}
\title{Create indicator species table}
\usage{
make_indicatortbl(indval)
}
\arguments{
\item{indval}{labdsv::indval Indicator value object}
}
\value{
knitr::kable table
}
\description{
Create indicator species table
}
