% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_predatortbl}
\alias{make_predatortbl}
\title{Make a table with predator occurrence and additional details}
\usage{
make_predatortbl(predators, stations)
}
\arguments{
\item{predators}{predator data aggregated to stations}

\item{stations}{station-level data}
}
\value{
knitr::knitr_kable
}
\description{
Make a table with predator occurrence and additional details
}
