% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_predpreytbl}
\alias{make_predpreytbl}
\title{Create predator-prey contingency table}
\usage{
make_predpreytbl(stations_clust)
}
\arguments{
\item{stations_clust}{Station-level cluster data}
}
\value{
knitr::kable table
}
\description{
Create predator-prey contingency table
}
