% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplemental.R
\name{make_tbl_effortice}
\alias{make_tbl_effortice}
\title{Make a table summarizing survey effort and ice conditions}
\usage{
make_tbl_effortice(stations_ice, station_effort, predators_stations)
}
\arguments{
\item{stations_ice}{Station-level ice data}

\item{station_effort}{Station-level effort data}

\item{predators_stations}{Station-level predator data}
}
\value{
knitr::knitr_kable
}
\description{
Make a table summarizing survey effort and ice conditions
}
