% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmds.R
\name{nmds_to_df}
\alias{nmds_to_df}
\title{Convert NMDS object to a data frame, including cluster and environmental data}
\usage{
nmds_to_df(nmds, clust, env)
}
\arguments{
\item{nmds}{NMDS object}

\item{clust}{Cluster data}

\item{env}{Environmental data}
}
\value{
NMDS object as data frame
}
\description{
Convert NMDS object to a data frame, including cluster and environmental data
}
