% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{project_bbox}
\alias{project_bbox}
\title{Project a bounding box}
\usage{
project_bbox(x)
}
\arguments{
\item{x}{sf::bbox}
}
\value{
sf::bbox x projected from lat/lon into Antarctic projection (see
ant_proj())
}
\description{
Project a bounding box
}
