% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{rasterize_by}
\alias{rasterize_by}
\title{Rasterize data (predator abundance or effort) by year or other category}
\usage{
rasterize_by(x_sf, field, by, fun, template)
}
\arguments{
\item{x_sf}{point data to rasterize}

\item{field}{field to rasterize}

\item{by}{grouping variable}

\item{fun}{rasterizing function}

\item{template}{raster template}
}
\value{
terra raster
}
\description{
Rasterize data (predator abundance or effort) by year or other category
}
