% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{read_obj}
\alias{read_obj}
\title{Read an object from derived_data/}
\usage{
read_obj(obj_name, type = c("r", "rast", "vect"))
}
\arguments{
\item{obj_name}{name of object}

\item{type}{whether object is a basic R object, raster, or vector. Raster and
vector data require additional processing.}
}
\value{
object read from file in analysis/data/derived_data/
}
\description{
Read an object from derived_data/
}
