% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoop.R
\name{read_zoop}
\alias{read_zoop}
\title{Read station (zooplankton) data}
\usage{
read_zoop(zoop_path)
}
\arguments{
\item{zoop_path}{path to StationZooplanktonPhysics.csv}
}
\value{
data frame with station-level data
}
\description{
Read station (zooplankton) data
}
