% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice.R
\name{seaice2contour}
\alias{seaice2contour}
\title{Create contours from seaice rasters}
\usage{
seaice2contour(seaice_rast, contour_level)
}
\arguments{
\item{seaice_rast}{terra::SpatRaster with sea ice concentrations}

\item{contour_level}{numeric vector with desired contour levels}
}
\value{
terra:SpatVector of sea ice contours
}
\description{
Create contours from seaice rasters
}
