% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice.R
\name{seaice2df}
\alias{seaice2df}
\title{Convert sea ice data from raster to data frame}
\usage{
seaice2df(seaice_rast)
}
\arguments{
\item{seaice_rast}{terra::SpatRaster with sea ice concentration data}
}
\value{
data frame
}
\description{
Convert sea ice data from raster to data frame
}
