% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmds.R
\name{shepard}
\alias{shepard}
\title{Create shepard diagram data}
\usage{
shepard(nmds)
}
\arguments{
\item{nmds}{nmds object (see vegan::metaMDS)}
}
\value{
tibble with diss, dist, and monotonically increasing nonparametric
fit
}
\description{
Create shepard diagram data
}
