% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmds.R
\name{stress}
\alias{stress}
\title{Calculate NMDS stress across numbers of axes}
\usage{
stress(dist, k)
}
\arguments{
\item{dist}{distance matrix}

\item{k}{number of axes (vectorized, e.g., 1:6 is ok)}
}
\value{
tibble with k and stress
}
\description{
Calculate NMDS stress across numbers of axes
}
