% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{read_obj}
\alias{read_obj}
\title{Read an object from derived_data/}
\usage{
read_obj(obj_name, rast = FALSE)
}
\arguments{
\item{obj_name}{name of object}

\item{rast}{whether object is a terra raster. false by default. requires
extra steps.}
}
\value{
object read from file in analysis/data/derived_data/
}
\description{
Read an object from derived_data/
}
