% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{save_obj}
\alias{save_obj}
\title{Save an object to derived_data/}
\usage{
save_obj(obj, obj_name, rast = FALSE)
}
\arguments{
\item{obj}{object to be saved}

\item{obj_name}{name of object}

\item{rast}{whether object is a terra raster. false by default. requires
extra steps.}
}
\value{
nothing, but saves object in analysis/data/derived_data/
}
\description{
Save an object to derived_data/
}
